third_party/WebKit/LayoutTests/fast/css/touch-action-parsing.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../resources/js-test.js"></script>
<style>
/* 
   Give some rules below something to override in order to test
   that they really are being parsed
*/ 
.defnone {
  touch-action: none;
}
</style>
</head>
<body>
<p id="description"></p>
<div class="test" id="default" expected="auto"></div>
<div class="test defnone" id="stylesheet-none" expected="none"></div>
<div class="test defnone" id="explicit-auto" style="touch-action: auto;" expected="auto"></div>
<div class="test" id="explicit-pan-x" style="touch-action: pan-x;" expected="pan-x"></div>
<div class="test" id="explicit-pan-y" style="touch-action: pan-y;" expected="pan-y"></div>
<div class="test" id="explicit-pan-x-pan-y" style="touch-action: pan-x pan-y;" expected="pan-x pan-y"></div>
<div class="test" id="explicit-pan-y-pan-x" style="touch-action: pan-y pan-x;" expected="pan-x pan-y"></div>
<div class="test" id="explicit-manipulation" style="touch-action: manipulation;" expected="manipulation"></div>
<div class="test" id="explicit-none" style="touch-action: none;" expected="none"></div>
<div class="test" id="explicit-invalid-1" style="touch-action: bogus;" expected="auto"></div>
<div class="test defnone" id="explicit-invalid-2" style="touch-action: auto pan-x;" expected="none"></div>
<div class="test" id="explicit-invalid-3" style="touch-action: pan-y none;" expected="auto"></div>
<div class="test" id="explicit-invalid-4" style="touch-action: pan-x pan-x;" expected="auto"></div>
<div class="test" id="explicit-invalid-5" style="touch-action: manipulation pan-x;" expected="auto"></div>
<div style="touch-action: none;">
  <div class="test" id="not-inherited" expected="auto"></div>
  <div class="test" id="inherit" style="touch-action: inherit;" expected="none"></div>
</div>
<div class="test defnone" id="initial" style="touch-action: initial;" expected="auto"></div>

<div id="console"></div>
<script>
description("Test the parsing and application of the touch-action property.");

var tests = document.querySelectorAll('.test');
var style;
for (var i = 0; i < tests.length; i++) {
  debug('Test case: ' + tests[i].id);
  style = window.getComputedStyle(tests[i]);
  shouldBeEqualToString('style.touchAction', tests[i].attributes.expected.value);
  debug('');
}

successfullyParsed = true;
</script>
</body>
</html>
