third_party/WebKit/LayoutTests/fast/css/touch-action-parsing-expected.txt

Test the parsing and application of the touch-action property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test case: default
PASS style.touchAction is "auto"

Test case: stylesheet-none
PASS style.touchAction is "none"

Test case: explicit-auto
PASS style.touchAction is "auto"

Test case: explicit-pan-x
PASS style.touchAction is "pan-x"

Test case: explicit-pan-y
PASS style.touchAction is "pan-y"

Test case: explicit-pan-x-pan-y
PASS style.touchAction is "pan-x pan-y"

Test case: explicit-pan-y-pan-x
PASS style.touchAction is "pan-x pan-y"

Test case: explicit-manipulation
PASS style.touchAction is "manipulation"

Test case: explicit-none
PASS style.touchAction is "none"

Test case: explicit-invalid-1
PASS style.touchAction is "auto"

Test case: explicit-invalid-2
PASS style.touchAction is "none"

Test case: explicit-invalid-3
PASS style.touchAction is "auto"

Test case: explicit-invalid-4
PASS style.touchAction is "auto"

Test case: explicit-invalid-5
PASS style.touchAction is "auto"

Test case: not-inherited
PASS style.touchAction is "auto"

Test case: inherit
PASS style.touchAction is "none"

Test case: initial
PASS style.touchAction is "auto"

PASS successfullyParsed is true

TEST COMPLETE
